requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
dataclasses-json==0.5.7
idna==3.3
lxml==4.9.1
markdownify==0.11.6
marshmallow==3.18.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-facebook-api==0.14.5
pyunsplash==1.0.0rc2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
typing-extensions==4.3.0
typing-inspect==0.8.0
urllib3==1.26.12
